+++ src/components/HomepageFeatures/styles.module.css
@@ -41,3 +41,33 @@
 .card:hover p {
   color: var(--ifm-color-primary);
 }
+
+.heroBanner {
+  padding: 3rem 0 1rem;
+  text-align: center;
+  position: relative;
+  overflow: hidden;
+}
+.heroBanner h1 {
+  font-weight: 600;
+  font-size: 2rem;
+  line-height: 2rem;
+  margin-bottom: 0.5rem;
+  font-family: Montserrat, sans-serif;
+  color: var(--ifm-color-primary);
+}
+.heroBanner p {
+  font-weight: 400;
+  font-family: Montserrat, sans-serif;
+  color: #727272;
+  font-size: 1.25rem;
+  font-weight: 600;
+  line-height: 1.25;
+  margin-bottom: 0;
+}
+
+@media screen and (max-width: 996px) {
+  .heroBanner {
+    padding: 2rem;
+  }
+}
