+++ src/pages/index.js
@@ -1,35 +1,15 @@
 import React from "react";
-import clsx from "clsx";
 import useDocusaurusContext from "@docusaurus/useDocusaurusContext";
 import Layout from "@theme/Layout";
 import HomepageFeatures from "@site/src/components/HomepageFeatures";
 
-import styles from "./index.module.css";
-
-function HomepageHeader() {
-  return (
-    <header className={clsx(styles.heroBanner)}>
-      <div className="container">
-        <h1 className={clsx("hero__title")}>
-          Welcome to the MultiversX (Elrond) docs!
-        </h1>
-        <p className={clsx("hero__subtitle")}>Choose your path you must.</p>
-      </div>
-    </header>
-  );
-}
-
 export default function Home() {
-  const { siteConfig } = useDocusaurusContext();
   return (
     <Layout
-      title={siteConfig.title}
-      description="Description will go into a meta tag in <head />"
+      title="Docs"
+      description="A highly scalable, fast and secure blockchain platform for distributed apps, enterprise use cases and the new internet economy."
     >
-      <HomepageHeader />
-      <main>
-        <HomepageFeatures />
-      </main>
+      <HomepageFeatures />
     </Layout>
   );
 }
